test: address review feedback on quick action button tests
P0: fix E2E task-creation assertion — use exact:true on getByText to
    avoid false-positive match against the toast substring; fix comment
    to correctly say "Recent Activity section" not "SpaceDetailPanel".

P1a: extract createSpaceViaRpc/deleteSpaceViaRpc into shared
     packages/e2e/tests/helpers/space-helpers.ts (following room-helpers
     pattern); update space-settings-crud.e2e.ts and space-task-creation
     to import from the shared helper instead of duplicating inline.

P1b: trim the redundant beforeEach inside the new unit-test describe
     block — the module-level beforeEach already resets all signals;
     the local override now only sets mockWorkflows = signal([]) with a
     comment explaining why (no workflows → dashboard fallback visible).

P2a: scope dialog-open assertions in E2E to the dialog element and use
     getByRole('heading') instead of fragile getByText().first() that
     could match the quick-action button behind the dialog.

P2b: add assertion in the Cancel test that no task pane opens after
     dismissal, explaining why (no task is selected so pane stays hidden).
